fix get responses to use modeled headers instead of explicit bytes
